+++ sknetwork/utils/__init__.py
@@ -2,7 +2,7 @@ from sknetwork.utils.formats import *
 from sknetwork.utils.kmeans import KMeans
 from sknetwork.utils.kneighbors import KNeighborsTransformer, FWKNeighborsTransformer
 from sknetwork.utils.projection_simplex import projection_simplex
-from sknetwork.utils.ward import Ward
+from sknetwork.utils.ward import WardDense
 
 
 class Bunch(dict):